+++ sdks/java/extensions/sql/zetasql/build.gradle
@@ -33,7 +33,6 @@ dependencies {
   compile enforcedPlatform(library.java.google_cloud_platform_libraries_bom)
   compile project(":sdks:java:core")
   compile project(":sdks:java:extensions:sql")
-  compile project(":sdks:java:extensions:sql:udf")
   compile library.java.vendored_calcite_1_20_0
   compile library.java.guava
   compile library.java.grpc_api
@@ -52,18 +51,5 @@ dependencies {
   testCompile library.java.mockito_core
   testCompile library.java.quickcheck_core
   testRuntimeClasspath library.java.slf4j_jdk14
-  testCompileOnly project(":sdks:java:extensions:sql:zetasql:udf-test-provider")
 }
 
-task emptyJar(type: Jar) {
-  archiveBaseName = "${project.archivesBaseName}-empty-jar"
-  from fileTree(dir: getTemporaryDir().createNewFile().toString())
-}
-
-test {
-  dependsOn emptyJar
-  // Pass jars used by Java UDF tests via system properties.
-  evaluationDependsOn(":sdks:java:extensions:sql:zetasql:udf-test-provider") // Needed to resolve jarPath.
-  systemProperty "beam.sql.udf.test.jar_path", project(":sdks:java:extensions:sql:zetasql:udf-test-provider").jarPath
-  systemProperty "beam.sql.udf.test.empty_jar_path", emptyJar.archivePath
-}
