+++ sdks/java/extensions/sql/zetasql/src/main/java/org/apache/beam/sdk/extensions/sql/zetasql/translation/UserFunctionDefinitions.java
@@ -21,7 +21,6 @@ import com.google.auto.value.AutoValue;
 import com.google.zetasql.resolvedast.ResolvedNode;
 import com.google.zetasql.resolvedast.ResolvedNodes;
 import java.util.List;
-import org.apache.beam.sdk.extensions.sql.udf.ScalarFn;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableMap;
 
 /** Holds user defined function definitions. */
@@ -37,8 +36,6 @@ public abstract class UserFunctionDefinitions {
    */
   public abstract ImmutableMap<List<String>, ResolvedNode> sqlTableValuedFunctions();
 
-  public abstract ImmutableMap<List<String>, ScalarFn> javaScalarFunctions();
-
   @AutoValue.Builder
   public abstract static class Builder {
     public abstract Builder setSqlScalarFunctions(
@@ -47,16 +44,12 @@ public abstract class UserFunctionDefinitions {
     public abstract Builder setSqlTableValuedFunctions(
         ImmutableMap<List<String>, ResolvedNode> sqlTableValuedFunctions);
 
-    public abstract Builder setJavaScalarFunctions(
-        ImmutableMap<List<String>, ScalarFn> javaScalarFunctions);
-
     public abstract UserFunctionDefinitions build();
   }
 
   public static Builder newBuilder() {
     return new AutoValue_UserFunctionDefinitions.Builder()
         .setSqlScalarFunctions(ImmutableMap.of())
-        .setSqlTableValuedFunctions(ImmutableMap.of())
-        .setJavaScalarFunctions(ImmutableMap.of());
+        .setSqlTableValuedFunctions(ImmutableMap.of());
   }
 }
