+++ settings.gradle
@@ -124,9 +124,9 @@ include ":sdks:java:extensions:sql:shell"
 include ":sdks:java:extensions:sql:hcatalog"
 include ":sdks:java:extensions:sql:datacatalog"
 include ":sdks:java:extensions:sql:zetasql"
-include ":sdks:java:extensions:sql:zetasql:udf-test-provider"
 include ":sdks:java:extensions:sql:expansion-service"
 include ":sdks:java:extensions:sql:udf"
+include ":sdks:java:extensions:sql:udf-test-provider"
 include ":sdks:java:extensions:zetasketch"
 include ":sdks:java:fn-execution"
 include ":sdks:java:harness"
